Big refactoring of rust impl + minor fixes
This refactors the Rust implementation to use methods instead of
functions for interfacing with the SliceStack struct.  It also allows
for runtime/cli selection of which execution path to take with regard
to safe and unsafe implementations.  No such selection of parallel path
yet.  Attempting to use more idiomatic Rust in main.rs.
Fixed some missing returns.
Commented out some C code for a more fare benchmark comparison (C impl
was doing more, and still is with the printfs()).
Next will be making logic.rs more idiomatic, using Option<i32> instead
of encoding failure into the return value (-1 usually).
As far as speed is concerned, the Rust and C impl are very similar with
C being less than %5 faster.  But stick Rayon in the mix and Rust is
over 2x faster with a minute amount of support code!